Change index link name and put at both ends
I changed the title of the link to "home" because I think it looks
better. I also added a duplicate to the bottom, for ease of access;
if bigger recipies are added I would want a link at the top and bottom,
so standardizing it now is good for consistency.

I also moved the top link above the title. I found that whenever I
opened one of the recipies my brain would read the link as if it was a
part of the recipie, because it was beneath the heading. Now people will
immediately understand that the link is not part of the recipie.