Add .Home-Title and underline
I made a class just for the title of the home page, because I wanted
to put an underline on it specifically.

This may be a time where an id would be better suited. I've chosen not
to use one because of my inexperience; I want to use classes so I can
find out how many times I end up needing a class I intended for one
thing for something else.