Organize css
I used comments to sort all of the code into three catagories, stuff
that affects everything, and stuff that affects home and recipies
respectively. I think this will make it easier to read going forward.
It should also be easy to modify later on if this method of sorting
becomes invalid